+++ equalsverifier-core/src/main/java/module-info.java
@@ -9,11 +9,8 @@ module nl.jqno.equalsverifier {
     requires org.objenesis;
 
     // Built-in prefab values
-    requires static com.google.common;
     requires static java.desktop;
     requires static java.naming;
     requires static java.rmi;
     requires static java.sql;
-    requires static javafx.base;
-    requires static org.joda.time;
 }